requirements.txt
# =====================================================
# Telco Network Optimization Suite - Python Dependencies
# =====================================================

# Core Streamlit
streamlit>=1.28.0

# Snowflake connectivity
snowflake-snowpark-python>=1.11.0
snowflake-connector-python>=3.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
plotly>=5.18.0
pydeck>=0.8.0
altair>=5.0.0
matplotlib>=3.7.0
branca>=0.6.0

# Geospatial
h3-py>=3.7.6

# Statistical analysis
scipy>=1.11.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

